Measure home advantage, and let a Wildcard be worth playing
Home advantage was not in the model at all. `strength_from_goal_model` divided
each venue by its own league baseline, which is arithmetically the same as
deleting the ratio home advantage lives in: all twenty clubs came back with
identical home and away multipliers, so `route_adjustment` graded a fixture on
the opponent alone. Measured before the fix: Arsenal 1.345 attack at home and
1.345 away; gameweek 14 away at Spurs and gameweek 34 at home to Spurs produced
byte-identical projections for all fifteen players.

Both venues now share one league baseline, so the fitted advantage survives, and
each club's own split is measured from its own fixtures on top of that. Dixon-
Coles fits one home advantage for the whole league; clubs plainly differ, so the
tilt is measured per club and shrunk toward the league over a nineteen-match
prior. Arsenal now read 1.487 at home against 1.203 away.

Fixture difficulty is continuous. Five buckets threw away most of what the route
model had measured — every Arsenal tie came out a 1 — so it is now `3 - k·log`
of the measured ease, to a tenth. Measured across the season: 41 distinct values
where there were 5, spread 97/206/121/158/64 across the bands with 14% clamped.
A promoted opponent is assumed soft rather than dropped, because reporting "no
fixture" for a tie that is being played is a lie; a club with no record of its
own still cannot be rated.

A Wildcard now re-plans the season around itself. The chained window solve is
long-termist by construction, so a squad must stay viable five weeks out and
never spends down. A segment that ends at the Wildcard has no week after it to
protect, which is the licence the chip actually buys. Solving the whole season
long-termist and then swapping in a rebuild — what it did before — paid the cost
and collected none of the benefit.

It still is not worth playing. With the greedy segment in place, both halves
measured worse than carrying on with one transfer a week, and the note now says
why: nothing in a projection built from completed-season rates ever breaks, so
there is nothing for a Wildcard to repair. That is a statement about the model,
not about the chip.

Triple Captain is ranked on the same blend the optimizer uses for the ordinary
armband rather than on the ceiling alone. A third copy adds exactly one more of
whatever the player scores, so on expected points the right pick is the highest
mean; the ceiling earns a third of a say because chasing a haul is the declared
strategy. Ranking on ceiling alone handed the chip to whoever was most volatile.

Season total: 2148.45 before any of this, 2181.36 now.

An analysis fetch failure names its status and request id, so the same failure
on production can be diagnosed from the page.

+++ python/fpl_andres/cli/publish_season_inputs.py
@@ -114,12 +114,12 @@ def main(argv: Sequence[str] | None = None) -> int:
     # "HUL (A)" per club per gameweek, so a solved card can name the opponent
     # rather than repeat the club whose shirt is already drawn beside the player.
     opponents: dict[str, list[list[str]]] = {}
-    # One to five per club per gameweek, the published difficulty of the tie.
-    ratings: dict[str, list[int | None]] = {}
+    # One to five per club per gameweek, to a tenth, the measured difficulty.
+    ratings: dict[str, list[float | None]] = {}
     for team_id, team in clubs.items():
         defensive: list[float] = []
         attacking: list[float] = []
-        difficulty: list[int | None] = []
+        difficulty: list[float | None] = []
         against: list[list[str]] = []
         for event in ordered:
             games = schedule.get((event, team_id), ())
